webcli general refactoring + adding first tests
