+++ jweb-console-webcli/package.json
@@ -7,6 +7,7 @@
     "dev": "vite",
     "build": "vite build --base=/console",
     "preview": "vite preview",
+    "test": "vitest",
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
@@ -16,7 +17,8 @@
     "svelte-check": "^2.10.3",
     "tslib": "^2.5.0",
     "typescript": "^4.9.3",
-    "vite": "^4.2.0"
+    "vite": "^4.2.0",
+    "vitest": "^0.30.1"
   },
   "dependencies": {
     "@codemirror/lang-javascript": "^6.1.4",
